+++ InstaBotPrototype/InstaBotLibrary/Instagram/IInstagramService.cs
@@ -2,7 +2,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using InstaBotLibrary.Posts;
 
 namespace InstaBotLibrary.Instagram
 {
